META-244: Fixed test AnalyticsDbCursor.reads_all_matched_data
META-244: Implemented analytics db cursor closure
VMS-13842: CloudStreamSocket::pollable() returned nullptr after CloudStreamSocket::shutdown()
VMS-13924. Server has crashed on Windows: c0000005 (Access violation)
VMS-13940: Server has crashed on Windows: c0000005 (Access violation
WEB-999: The 'Playing position' is displayed incorrectly.

<==== Mercurial data ====>
Branch: vms_4.0

+++ vms/server/nx_vms_server/src/plugins/resource/flir/eip_cip.cpp
@@ -46,7 +46,7 @@ MessageRouterResponse MessageRouterResponse::decode(const QByteArray& buf)
         response.additionalStatus.append(tmp.get(), response.sizeOfAdditionalStatus);
     }
 
-    size_t sizeOfData = buf.size()
+    int sizeOfData = buf.size()
         - sizeof(response.serviceCode)
         - sizeof(response.reserved)
         - sizeof(response.generalStatus)
